+++ arrus/core/devices/us4r/us4oem/Us4OEMDescriptorFactory.h
@@ -23,9 +23,9 @@ public:
                 version, // Us4OEM version
                 32, // RX channels
                 20e-6f,  // min. RX time,
-                5e-6, // RX time epsilon,
-                35e-6, // TX parameters reprogramming time,
-                65e6, // Sampling frequency [Hz]
+                5e-6f, // RX time epsilon,
+                35e-6f, // TX parameters reprogramming time,
+                65e6f, // Sampling frequency [Hz]
                 1ull << 32u, // DDR memory size [B]
                 1ull << (14+12), // Max transfer size [B]
                 0.5f,  // number of TX periods resolution
@@ -35,7 +35,7 @@ public:
                         arrus::ops::us4r::TxLimits {
                             Interval<float>{minFrequency, maxFrequency},  // Frequency
                             Interval<float>{0.0f, 16.96e-6f}, // delay
-                            Interval<float>{0.0f, 32.0f/10e6}, // pulse length,
+                            Interval<float>{0.0f, (float)(32.0f/10e6)}, // pulse length,
                             Interval<Voltage>{5, 90}
                         },
                         arrus::ops::us4r::RxLimits {
@@ -53,9 +53,9 @@ public:
                 version, // us4OEM version
                 32, // RX channels
                 20e-6f,  // min. RX time,
-                5e-6, // RX time epsilon,
-                35e-6, // TX parameters reprogramming time,
-                65e6, // Sampling frequency [Hz]
+                5e-6f, // RX time epsilon,
+                35e-6f, // TX parameters reprogramming time,
+                65e6f, // Sampling frequency [Hz]
                 1ull << 32u, // DDR memory size [B]
                 1ull << (14+12), // Max transfer size [B]
                 0.5f,  // number of TX periods resolution
@@ -65,7 +65,7 @@ public:
                         arrus::ops::us4r::TxLimits {
                             Interval<float>{minFrequency, maxFrequency},  // Frequency
                             Interval<float>{0.0f, 16.96e-6f}, // delay
-                            Interval<float>{0.0f, 32.0f/10e6}, // pulse length,
+                            Interval<float>{0.0f, (float)(32.0f/10e6f)}, // pulse length,
                             Interval<Voltage>{5, 90}
                         },
                         arrus::ops::us4r::RxLimits {
