+++ arrus/core/devices/us4r/us4oem/Us4OEMImpl.h
@@ -160,10 +160,10 @@ private:
                        const std::optional<ops::us4r::DigitalDownConversion> &ddc,
                        const std::vector<arrus::framework::NdArray> &txDelays,
                        const Us4OEMRxMappingRegister &rxMappingRegister);
-    size_t scheduleReceiveDDC(size_t outputAddress, uint16 startSample, uint16 endSample, uint16 entryId,
+    size_t scheduleReceiveDDC(size_t outputAddress, uint32 startSample, uint32 endSample, uint16 entryId,
                               const us4r::TxRxParameters &op, uint16 rxMapId,
                               const std::optional<ops::us4r::DigitalDownConversion> &ddc);
-    size_t scheduleReceiveRF(size_t outputAddress, uint16 startSample, uint16 endSample, uint16 entryId,
+    size_t scheduleReceiveRF(size_t outputAddress, uint32 startSample, uint32 endSample, uint16 entryId,
                              const us4r::TxRxParameters &op, uint16 rxMapId);
     Us4OEMBuffer uploadAcquisition(const us4r::TxParametersSequenceColl &sequences, uint16 rxBufferSize,
                                    const std::optional<ops::us4r::DigitalDownConversion> &ddc,
